Split js code into  modules
